added simple toolbar as component
