+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-viewerbase",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "Cornerstone medical image viewport component for React",
   "author": "Cornerstone Contributors",
   "license": "MIT",
