+++ src/index.js
@@ -1,13 +1,18 @@
-import LayoutChooser from './LayoutChooser/LayoutChooser';
+import LayoutChooser from './LayoutChooser/LayoutChooser.js';
 import LayoutButton from './LayoutChooser/LayoutButton.js';
 import ToolbarSection from './viewer/ToolbarSection.js';
 import StudyBrowser from './studyBrowser/StudyBrowser.js';
 import ThumbnailEntry from './studyBrowser/ThumbnailEntry.js';
+import Toolbar from './viewer/Toolbar.js';
+import PlayClipButton from './viewer/PlayClipButton.js';
 
 export {
+  LayoutChooser,
   LayoutButton,
   ToolbarSection,
   StudyBrowser,
   ThumbnailEntry,
-  LayoutChooser
+  Toolbar,
+  PlayClipButton
 };
+//
