+++ src/viewer/SimpleToolbarButton.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 export function SimpleToolbarButton(props) {
